Completed react-validated-input-component challenge
